directory restr
