+++ components/logoComposition.tsx
