+++ package.json
@@ -12,9 +12,13 @@
   },
   "dependencies": {
     "flowbite-react": "^0.12.4",
+    "i18next": "^25.6.3",
+    "motion": "^12.23.24",
     "next": "^15.4.2",
+    "next-i18next": "^15.4.2",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-i18next": "^16.3.4",
     "react-icons": "^5.5.0"
   },
   "devDependencies": {
