+++ pages/_app.tsx
@@ -1,0 +1,8 @@
+import "@/styles/globals.css";
+import type { AppProps } from "next/app";
+import { appWithTranslation } from "next-i18next";
+
+function App({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />;
+}
+export default appWithTranslation(App /*, nextI18NextConfig */);
