+++ tsconfig.json
@@ -27,7 +27,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "app/page.jsx"
+    "pages/page.jsx"
   ],
   "exclude": ["node_modules"]
 }
